refactor QSP/GQSP helper code
