+++ qualtran/bloqs/qsp/generalized_qsp_test.py
@@ -12,20 +12,23 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 from functools import cached_property
-from typing import Dict, Optional, Sequence, Union
+from typing import Dict, Optional, Sequence
 
 import cirq
 import numpy as np
 import pytest
 import sympy
 from attrs import define
-from numpy.polynomial import Polynomial
-from numpy.typing import NDArray
 
 from qualtran import Bloq, bloq_example, Controlled, CtrlSpec, GateWithRegisters
 from qualtran.bloqs.basic_gates.su2_rotation import SU2RotationGate
 from qualtran.bloqs.for_testing.atom import TestGWRAtom
 from qualtran.bloqs.for_testing.matrix_gate import MatrixGate
+from qualtran.linalg.polynomial.qsp_util import (
+    check_polynomial_pair_on_random_points_on_unit_circle,
+    evaluate_polynomial_of_matrix,
+    random_qsp_polynomial,
+)
 from qualtran.bloqs.qsp.generalized_qsp import (
     _gqsp,
     _gqsp_with_large_negative_power,
@@ -34,7 +37,6 @@ from qualtran.bloqs.qsp.generalized_qsp import (
     GeneralizedQSP,
     qsp_complementary_polynomial,
     qsp_phase_factors,
-    scale_down_to_qsp_polynomial,
 )
 from qualtran.bloqs.qubitization.qubitization_walk_operator_test import (
     get_walk_operator_for_1d_ising_model,
@@ -47,38 +49,6 @@ def test_gqsp_example(bloq_autotester):
     bloq_autotester(_gqsp)
     bloq_autotester(_gqsp_with_negative_power)
     bloq_autotester(_gqsp_with_large_negative_power)
-
-
-def assert_angles_almost_equal(
-    actual: Union[float, Sequence[float]], desired: Union[float, Sequence[float]]
-):
-    """Helper to check if two angle sequences are equal (up to multiples of 2*pi)"""
-    np.testing.assert_almost_equal(np.exp(np.array(actual) * 1j), np.exp(np.array(desired) * 1j))
-
-
-def check_polynomial_pair_on_random_points_on_unit_circle(
-    P: Union[Sequence[complex], Polynomial, Shaped],
-    Q: Union[Sequence[complex], Polynomial, Shaped],
-    *,
-    random_state: np.random.RandomState,
-    rtol: float = 1e-7,
-    n_points: int = 1000,
-):
-    P = Polynomial(P)
-    Q = Polynomial(Q)
-
-    for _ in range(n_points):
-        z = np.exp(random_state.random() * np.pi * 2j)
-        np.testing.assert_allclose(np.abs(P(z)) ** 2 + np.abs(Q(z)) ** 2, 1, rtol=rtol)
-
-
-def random_qsp_polynomial(
-    degree: int, *, random_state: np.random.RandomState, only_real_coeffs=False
-) -> Sequence[complex]:
-    poly = random_state.random(size=degree) / degree
-    if not only_real_coeffs:
-        poly = poly * np.exp(random_state.random(size=degree) * np.pi * 2j)
-    return list(poly)
 
 
 @pytest.mark.parametrize("degree", [3, 4, 5, 10, 40, pytest.param(100, marks=pytest.mark.slow)])
@@ -100,26 +70,6 @@ def test_real_polynomial_has_real_complementary_polynomial(degree: int):
         Q = qsp_complementary_polynomial(P, verify=True)
         Q = np.around(Q, decimals=8)
         assert np.isreal(Q).all()
-
-
-def evaluate_polynomial_of_matrix(
-    P: Sequence[complex], U: NDArray, *, negative_power: int = 0
-) -> NDArray:
-    assert U.ndim == 2 and U.shape[0] == U.shape[1]
-
-    pow_U = np.linalg.matrix_power(U.conj().T, negative_power)
-    result = np.zeros(U.shape, dtype=U.dtype)
-
-    for c in P:
-        result += pow_U * c
-        pow_U = pow_U @ U
-
-    return result
-
-
-def assert_matrices_almost_equal(A: NDArray, B: NDArray, *, atol: float = 1e-5):
-    assert A.shape == B.shape
-    assert np.linalg.norm(A - B) <= atol
 
 
 def verify_generalized_qsp(
@@ -333,7 +283,7 @@ def test_generalized_real_qsp_with_symbolic_signal_matrix(degree: int):
 @pytest.mark.parametrize("t", [2, 5, 7])
 @pytest.mark.parametrize("precision", [1e-4, 1e-7, 1e-9])
 def test_complementary_polynomials_for_jacobi_anger_approximations(t: float, precision: float):
-    from qualtran.linalg.jacobi_anger_approximations import (
+    from qualtran.linalg.polynomial.jacobi_anger_approximations import (
         approx_exp_cos_by_jacobi_anger,
         degree_jacobi_anger_approximation,
     )
