+++ qualtran/linalg/polynomial/jacobi_anger_approximations.py
